Add watchdog.
There are too many things in the internals of the USB connection
to the NFC device that are outside of our control, that it is best
to have a watchdog exiting the process when nothing is happening
for a while.

The InitiatorListPassiveTargets() should return regularly, which
is when we trigger the WatchDog.